Add files via upload
Landing Page for OctaNet Services Pvt Ltd - Web Development Internship

Project Overview This repository contains the code for my first task as a Web Development Intern at OctaNet Services Pvt Ltd. The project is a simple landing page designed using HTML and CSS.The landing page is divided into three sections:

Home (Landing Page 01): A navigation bar with the home , about and a contact button to navigate to different sections and register button. The main content includes the courses that we provide.

About (Page 02): This section details the educational environment we provide.

Contact (Page 03): A contact section with the details plus a form for any enquiry the learner wants to ask.

Technologies Used- HTML: For structuring the content of the landing page. CSS: For styling the landing page, including layout, colors, and fonts.